Migrate codebase to PSR-4 namespace
https://getcomposer.org/doc/04-schema.md#autoload
