+++ src/CeleryException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Celery;
+
+/**
+ * General exception class
+ * @package celery-php
+ */
+class CeleryException extends \Exception
+{
+}
