+++ src/CeleryPublishException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Celery;
+
+/**
+ * Emited by CeleryAbstract::PostTask() connection failures etc
+ * @package celery-php
+ */
+class CeleryPublishException extends CeleryException
+{
+}
