+++ src/CeleryTimeoutException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Celery;
+
+/**
+ * Emited by AsyncResult::get() on timeout
+ * @package celery-php
+ */
+class CeleryTimeoutException extends CeleryException
+{
+}
