+++ test.php
@@ -34,9 +34,9 @@
  * gdr@gdr.name
  */
 
-require_once('celery.php');
+require_once 'vendor/autoload.php';
 
-$c = new Celery('localhost', 'gdr', 'test', 'wutka', 'celery', 'celery', 5672, 'php-amqplib');
+$c = new \Celery\Celery('localhost', 'gdr', 'test', 'wutka', 'celery', 'celery', 5672, 'php-amqplib');
 #$result = $c->PostTask('tasks.add', array(2,2));
 $result = $c->PostTask('tasks.delayed', []);
 #$result = $c->PostTask('tasks.fail', array());
